mediguide-backend/requirements.txt
# MediGuide Backend Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase==2.3.0
postgrest==0.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
pydantic-settings==2.1.0

# AI/ML
openai==1.12.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Image processing
opencv-python-headless==4.9.0.80

# Utilities
python-dateutil==2.8.2
pydantic==2.5.3

# Optional: For production background tasks
# celery==5.3.4
# redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
